Merge to master

+++ src/actions/team.ts
@@ -1,62 +1,70 @@
-import { ThunkAction } from '@reduxjs/toolkit';
-import { processResponse, jsonHeaders } from '../core/ApiClient';
-import { Action, State, Team } from '../interfaces';
+import { ThunkAction } from "@reduxjs/toolkit";
+import { processResponse, jsonHeaders } from "../core/ApiClient";
+import { Action, State, Team } from "../interfaces";
 
 export function deleteTeam(): Action {
   return {
-    type: "DELETE_TEAM"
+    type: "DELETE_TEAM",
   };
 }
 
 export function teamDeleted(): Action {
   return {
-    type: "TEAM_DELETED"
+    type: "TEAM_DELETED",
   };
 }
 
 export function removeTeam(): ThunkAction<void, State, unknown, Action> {
   return (dispatch, getState) => {
     const state = getState();
+    if (!state.team) {
+      return Promise.reject(new Error("No team selected"));
+    }
     const teamId = state.team.id;
     const host = state.host;
     dispatch(deleteTeam());
     return fetch(`//${host}/api/teams/${teamId}`, {
-      method: 'delete',
-      credentials: 'include',
-      headers: jsonHeaders
+      method: "delete",
+      credentials: "include",
+      headers: jsonHeaders,
     })
-      .then(response => processResponse(response, dispatch))
+      .then((response) => processResponse(response, dispatch))
       .then(() => dispatch(teamDeleted()));
   };
 }
 
-export function patchTeam(obj: Team): Action {
+export function patchTeam(obj: Partial<Team>): Action {
   return {
     type: "PATCH_TEAM",
-    team: obj
+    team: obj,
   };
 }
 
 export function teamPatched(json: Team): Action {
   return {
     type: "TEAM_PATCHED",
-    team: json
+    team: json,
   };
 }
 
-export function updateTeam(payload: Team): ThunkAction<void, State, unknown, Action> {
+export function updateTeam(
+  payload: Partial<Team>
+): ThunkAction<Promise<Action>, State, unknown, Action> {
   return (dispatch, getState) => {
     const state = getState();
+    if (!state.team) {
+      return Promise.reject(new Error("No team selected"));
+    }
     const teamId = state.team.id;
     const host = state.host;
     dispatch(patchTeam(payload));
     return fetch(`//${host}/api/teams/${teamId}`, {
-      method: 'PATCH',
-      credentials: 'include',
+      method: "PATCH",
+      credentials: "include",
       headers: jsonHeaders,
-      body: JSON.stringify(payload)
+      body: JSON.stringify(payload),
     })
-      .then(response => processResponse(response, dispatch))
-      .then(json => dispatch(teamPatched(json)));
+      .then((response) => processResponse(response, dispatch))
+      .then((json) => dispatch(teamPatched(json)));
   };
 }
